Moving eventqueue into new event directory.
